Factor our the common channel provider and the proxy channel provider as singletons
